# Conflicts:
#	jmp/mp-db-tables/src/main/java/com/vpu/mp/db/shop/tables/PrescriptionItem.java

+++ jmp/mp-common/src/main/java/com/vpu/mp/service/pojo/shop/prescription/PrescriptionItemInfoVo.java
@@ -52,7 +52,7 @@ public class PrescriptionItemInfoVo {
     /**
      * 总取药数量
      */
-    private Integer   dragSumNum;
+    private Double   dragSumNum;
     /**
      * 总取药单位
      */
